bcmdhd: workaround for driver compiled as a module
BCMDHD driver on several Samsung and few Sony devices is compiled as a
module rather than being built into the kernel. Hence, the wlan0
interface
is removed when the WiFi is turned off and the driver is unloaded. This
causes the WiFi HAL to crash when the device is booted with WiFi turned
off
since with the absence of wlan0 interface, the WiFi interfaces in the
HAL
are not initialized.

As a workaround, we check if the driver is loaded and if not, we
temporarily load
and unload the WiFi driver inorder to query the interfaces and
initialize them.
This atleast prevents the system from crashing and rebooting when the
device is suspended.

https://github.com/ArchiDroid/android_hardware_broadcom_wlan/commit/f52904cb41f38e8bb2d3048253f93b29272b0544